refactor : RacingGameRunner 를 RacingGameController 로 rename하고 게임 동작 로직을 모음
